rename: service to checkpoint_service
